refactor(pkg/oomstore): move  batch functions to import_batch
