+++ pkg/oomstore/import.go
@@ -1,10 +1,8 @@
 package oomstore
 
 import (
-	"bufio"
 	"context"
 	"os"
-	"time"
 
 	"github.com/oom-ai/oomstore/pkg/errdefs"
 	"github.com/spf13/cast"
@@ -55,88 +53,40 @@ func (s *OomStore) Import(ctx context.Context, opt types.ImportOpt) (int, error)
 	}
 }
 
-func (s *OomStore) csvReaderImportBatch(ctx context.Context, opt *importOpt, dataSource *types.CsvReaderDataSource) (int, error) {
-	//make sure csv data source has all defined columns
-	reader := bufio.NewReader(dataSource.Reader)
-	source := &offline.CSVSource{
-		Reader:    reader,
-		Delimiter: dataSource.Delimiter,
-	}
-	// read header
-	columnNames := append([]string{opt.entityName}, opt.features.Names()...)
-	header, err := readHeader(source, columnNames)
-	if err != nil {
-		return 0, err
-	}
-
-	newRevisionID, err := s.createRevision(ctx, metadata.CreateRevisionOpt{
-		Revision:      time.Now().UnixMilli(),
-		GroupID:       opt.group.ID,
-		SnapshotTable: nil,
-		Description:   opt.Description,
-		Anchored:      opt.Revision != nil,
-	})
-	if err != nil {
-		return 0, err
-	}
-	snapshotTableName := dbutil.OfflineBatchSnapshotTableName(opt.group.ID, int64(newRevisionID))
-	revision, err := s.offline.Import(ctx, offline.ImportOpt{
-		EntityName:        opt.entityName,
-		Features:          opt.features,
-		Header:            cast.ToStringSlice(header),
-		Revision:          opt.Revision,
-		SnapshotTableName: snapshotTableName,
-		Source:            source,
-	})
-	if err != nil {
-		return 0, err
-	}
-
-	if opt.Revision != nil {
-		revision = *opt.Revision
-	}
-	if err = s.metadata.UpdateRevision(ctx, metadata.UpdateRevisionOpt{
-		RevisionID:       newRevisionID,
-		NewRevision:      &revision,
-		NewSnapshotTable: &snapshotTableName,
-	}); err != nil {
-		return 0, err
-	}
-
-	// TODO: clean up revision and data_table if import failed
-
-	return newRevisionID, nil
+type importOpt struct {
+	*types.ImportOpt
+	entityName string
+	group      *types.Group
+	features   types.FeatureList
 }
 
-func (s *OomStore) tableLinkImportBatch(ctx context.Context, opt *importOpt, dataSource *types.TableLinkDataSource) (int, error) {
-	// Make sure all features existing with correct value type
-	tableSchema, err := s.offline.TableSchema(ctx, offline.TableSchemaOpt{
-		TableName: dataSource.TableName,
-	})
+func (s *OomStore) parseImportOpt(ctx context.Context, opt types.ImportOpt) (*importOpt, error) {
+	group, err := s.metadata.GetGroupByName(ctx, opt.GroupName)
 	if err != nil {
-		return 0, err
-	}
-	if err = validateTableSchema(tableSchema, opt.features); err != nil {
-		return 0, err
-	}
-	var revision int64
-	if opt.Revision == nil {
-		revision = time.Now().UnixMilli()
-	} else {
-		revision = *opt.Revision
-	}
-	newRevisionID, err := s.createRevision(ctx, metadata.CreateRevisionOpt{
-		Revision:      revision,
-		GroupID:       opt.group.ID,
-		SnapshotTable: &dataSource.TableName,
-		Description:   opt.Description,
-		Anchored:      opt.Revision != nil,
-	})
-	if err != nil {
-		return 0, err
+		return nil, err
 	}
 
-	return newRevisionID, nil
+	features, err := s.metadata.ListFeature(ctx, metadata.ListFeatureOpt{
+		GroupID: &group.ID,
+	})
+	if err != nil {
+		return nil, err
+	}
+	if features == nil {
+		err = errdefs.Errorf("no features under group: %s", opt.GroupName)
+		return nil, err
+	}
+
+	entity := group.Entity
+	if entity == nil {
+		return nil, errdefs.Errorf("no entity found by group: %s", opt.GroupName)
+	}
+	return &importOpt{
+		ImportOpt:  &opt,
+		entityName: entity.Name,
+		group:      group,
+		features:   features,
+	}, nil
 }
 
 func validateTableSchema(schema *types.DataTableSchema, features types.FeatureList) error {
@@ -205,40 +155,4 @@ func stringSliceEqual(a, b []string) bool {
 		}
 	}
 	return true
-}
-
-func (s *OomStore) parseImportOpt(ctx context.Context, opt types.ImportOpt) (*importOpt, error) {
-	group, err := s.metadata.GetGroupByName(ctx, opt.GroupName)
-	if err != nil {
-		return nil, err
-	}
-
-	features, err := s.metadata.ListFeature(ctx, metadata.ListFeatureOpt{
-		GroupID: &group.ID,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if features == nil {
-		err = errdefs.Errorf("no features under group: %s", opt.GroupName)
-		return nil, err
-	}
-
-	entity := group.Entity
-	if entity == nil {
-		return nil, errdefs.Errorf("no entity found by group: %s", opt.GroupName)
-	}
-	return &importOpt{
-		ImportOpt:  &opt,
-		entityName: entity.Name,
-		group:      group,
-		features:   features,
-	}, nil
-}
-
-type importOpt struct {
-	*types.ImportOpt
-	entityName string
-	group      *types.Group
-	features   types.FeatureList
 }
